Fix HTTPS error and secret-authentication audience parameter (#1)
This pull request fixes two issues we discovered while trying to use iotedge-aad. The first was that authentication requests would fail with the hyper error Invalid scheme for Http. This issue is fixed by dac9bd9. The second issue was that in spite of iotedge_aad::Auth being initialized with, say, the resource https://vault.azure.net, we would still be faced with an Invalid audience error. This issue is fixed by 804f471.